papertrail/requirements.txt
django-jsonfield>=0.8.11
django-apptemplates>=1.0
